[C8] Split sketched_features.go per feature (#2239)
model/feature/sketched_features.go was 862 LOC, 1.7x the 500-line ceiling. It
held the revolve, coil and rib sketch-driven features. Split per feature; each
file is now under the ceiling:

- revolve.go (200): the revolve feature (definition, Recompute, axis
  resolution) and the RevolveFeatures adder collection.
- revolve_build.go (305): the revolve tool geometry builders — the analytic
  torus/sphere fast paths, the meridian vertices, and the general revolved-
  section sweep (split out to keep revolve.go well under the ceiling).
- coil.go (131): the coil feature + CoilFeatures collection.
- rib.go (233): the rib feature + RibFeatures collection.
- sketched_features.go (57): the shared sketch-profile binding helpers
  (resolveSingleProfile, featOr, ensureCCW2).

Behaviour-preserving: split programmatically by receiver-qualified
declaration, verified every non-comment code line is byte-identical (only
import redistribution differs); imports resolved per file with goimports.
Feature package builds, archguard passes, lint 0. Net delta of recognizers /
tolerance constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2239.

(cherry picked from commit 628b5c6f6e0afc2440e96efe90cbfa715a15053a)
